third_party/blink/web_tests/wpt_internal/storage/buckets/detached-iframe.https.html

<!doctype html>
<meta charset=utf-8>
<title>Storage Buckets API on detached iframe</title>
<link rel='help' href=''>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
'use strict';

promise_test(async testCase => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);
  const bucketManager = iframe.contentWindow.navigator.storageBuckets;
  await bucketManager.open('bucket');

  let bucketKeys = await bucketManager.keys();
  assert_equals(bucketKeys.length, 1);

  iframe.remove();

  // Calling open() from a detached iframe should fail but not crash.
  assert_equals(bucketManager.open("bucket"), undefined);
}, 'Verify open() on detached iframe returns an error');

promise_test(async testCase => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);
  const bucketManager = iframe.contentWindow.navigator.storageBuckets;
  await bucketManager.open('bucket');

  let bucketKeys = await bucketManager.keys();
  assert_equals(bucketKeys.length, 1);

  iframe.remove();

  // Calling keys() from a detached iframe should fail but not crash.
  assert_equals(bucketManager.keys(), undefined);
}, 'Verify keys() on detached iframe returns an error');

promise_test(async testCase => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);
  const bucketManager = iframe.contentWindow.navigator.storageBuckets;
  await bucketManager.open('bucket');

  iframe.remove();

  // Calling delete() from a detached iframe should fail but not crash.
  assert_equals(bucketManager.delete('bucket)'), undefined);
}, 'Verify delete() on detached iframe returns an error');

</script>
</body>
